refactor: Phase 2 Command Preprocessor - Extract to ids/resolver.go (#47)
- Extract CommandPreprocessor to ids/resolver.go with clean interfaces
- Create IDResolver interface for ID resolution
- Create FieldInfo interface for dimension field knowledge
- Extract FieldResolver to types/field_resolver.go
- Move command types to separate commands.go file
- Update command_preprocessor.go to be a thin wrapper
- Store now implements IDResolver interface
- All tests passing, no functionality changed
